requirement.txt
tensorflow==2.16.1
flask
Pillow
numpy